开发 Specs 后端界面；开发 Specs 后台功能；

+++ system/plugins/specs/core/FieldTypeDefinitions/FloatType.php
@@ -11,7 +11,7 @@ class FloatType extends DefinitionBase
      *
      * @var string
      */
-    protected $id = 'integer';
+    protected $id = 'float';
 
     /**
      * 字段类型标题
